WiproAssignments/javaAssignments/javaAssignments.day8.part2/src/main/java/javaAssignments/day8/part2/task5/MathematicalOperations.java
/* Task 2: Extend the above JUnit tests to use @Before, @After, @BeforeClass, and @AfterClass annotations to manage test setup and teardown. */

package javaAssignments.day8.part2.task5;

public class MathematicalOperations {

    public int add(int a, int b) {
        return a + b;
    }

    public int subtract(int a, int b) {
        return a - b;
    }

    public int multiply(int a, int b) {
        return a * b;
    }

    public double divide(int dividend, int divisor) {
        if (divisor == 0) {
            throw new IllegalArgumentException("Cannot divide by zero");
        }
        return (double) dividend / divisor;
    }
}
